requirements.txt
langchain
openai
langchain-openai
chromadb
gradio
python-dotenv
scikit-learn
plotly
langchain-community
